+++ starlite/controller.py
@@ -12,9 +12,9 @@ from starlite.types import (
 from starlite.utils import normalize_path
 
 if TYPE_CHECKING:  # pragma: no cover
+    from starlite import Router
     from starlite.handlers import BaseRouteHandler
     from starlite.provide import Provide
-    from starlite.routing import Router
 
 
 class Controller:
