+++ starlite/handlers/base.py
@@ -23,8 +23,8 @@ from starlite.types import Guard
 from starlite.utils import normalize_path
 
 if TYPE_CHECKING:  # pragma: no cover
+    from starlite import Router
     from starlite.controller import Controller
-    from starlite.routing import Router
 
 
 class BaseRouteHandler:
