+++ tests/test_router_registration.py
@@ -4,6 +4,7 @@ from starlite import (
     Controller,
     HttpMethod,
     ImproperlyConfiguredException,
+    Router,
     WebSocket,
     get,
     patch,
@@ -11,7 +12,6 @@ from starlite import (
 )
 from starlite import route as route_decorator
 from starlite import websocket
-from starlite.routing import Router
 
 
 class MyController(Controller):
